This commit reworks the logging functionality behind Aspen. 1. I removed my homemade logging functionality and wrapped a new Log object around boost's logging functionality. 2. This new logger highlighted some issues; for example, there was a script severity level which needed to be removed. Script modules should be able to log critical and others without being relegated to one specific severity level. 3. We bumped to requiring c++14, mostly for make_unique. This is supported by all modern compilers.
This change means that current code that called World->WriteLog will just now need to call WriteLog. There are two versions of this method; the first, taking simply a string allows you to write a single message, and all of these messages are logged as severity info by default. The second now takes a scoped enum, and you will need to pass in a severity level, which can be found in log.h.

This system was designed using the PIMPL pattern; all boost logic is wrapped in log.cpp and the log.h file does not need to know about it. This feels slightly messy, but I want to make sure that nothing of the implementation needs to leak out. There are still some various issues with this code which I will probably be cleaning up in the next commit. Mostly, we need a clean way to dispose of the Log object itself. Calling Release or manually deleting it doesn't properly dispose of it should anything return.